Fix support for `include()` within `multiurl()`.
The try/except block was not copied across from `RegexURLResolver`, so while
looping through and attempting to match patterns, a `Resolver404` could be
raised when `pattern` was itself a `RegexURLResolver` (returned by
`include()`) and none of it's patterns matched. This exception would bubble
up and the rest of the patterns would not be tested.